separated the processes of generating the menu and sending messages
